restructure tests. Attempt setup of browser tests on travis
