+++ .gitignore
@@ -12,3 +12,6 @@ node_modules
 *.map
 
 .env
+
+*.log
+*.png
