+++ .travis.yml
@@ -1,13 +1,18 @@
+sudo: required
 language: php
+dist: trusty
 
 php:
   - 7.1
   - 7.2
 
 before_script:
+  - export DISPLAY=:99.0
+  - sh -e /etc/init.d/xvfb start
+  - ./vendor/laravel/dusk/bin/chromedriver-linux &
   - cp .env.travis .env
-  - composer self-update
-  - composer install --no-interaction
+  - travis_retry composer install --no-interaction
 
 script:
-  - vendor/bin/phpunit tests/Unit/PackageTestCase.php
+  - vendor/bin/phpunit tests/Unit/PackageTestCases.php
+  - vendor/bin/phpunit tests/Browser/BrowserTestCases.php
