+++ phpunit.xml
@@ -9,12 +9,12 @@
          processIsolation="false"
          stopOnFailure="false">
     <testsuites>
-        <testsuite name="Feature">
-            <directory suffix="Test.php">./tests/Feature</directory>
+        <testsuite name="browser">
+            <directory suffix=".php">./tests/Browser</directory>
         </testsuite>
 
-        <testsuite name="Unit">
-            <directory suffix="Test.php">./tests/Unit</directory>
+        <testsuite name="unit">
+            <directory suffix=".php">./tests/Unit</directory>
         </testsuite>
     </testsuites>
     <filter>
